refactor: move materials out to reduce noise & fix types
